+++ superset-frontend/plugins/plugin-chart-echarts/src/BigNumber/src/BigNumber/images/thumbnailLarge.png
