+++ superset-frontend/plugins/plugin-chart-echarts/src/BigNumber/src/BigNumberTotal/images/thumbnailLarge.png
