+++ superset-frontend/plugins/plugin-chart-echarts/src/BigNumber/src/sharedControls.ts
